backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.23
alembic>=1.12.1
httpx>=0.27.0
pydantic>=2.9.0
pydantic-settings>=2.4.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
websockets>=12.0
redis>=5.0.1
celery>=5.3.4
pytest>=7.4.3
pytest-asyncio>=0.21.1
